// back/generated/opendart/rest/api.ts
export const OpendartCorpCodeRegExp = new RegExp("^[0-9]{8}$")

export const OpendartCorpCode = /*#__PURE__*/ zod
  .string()
  .check(/*#__PURE__*/ zod.regex(OpendartCorpCodeRegExp))

export type OpendartCorpCode = zod.input<typeof OpendartCorpCode>
export type OpendartCorpCodeOutput = zod.output<typeof OpendartCorpCode>

export const OpendartErrorResponse = /*#__PURE__*/ zod.object({
  status: /*#__PURE__*/ zod.string().check(/*#__PURE__*/ zod.minLength(1)),
  message: /*#__PURE__*/ zod.optional(/*#__PURE__*/ zod.string()),
})

export type OpendartErrorResponse = zod.input<typeof OpendartErrorResponse>
export type OpendartErrorResponseOutput = zod.output<
  typeof OpendartErrorResponse
>

export const opendartNullableStockCodeOneRegExp = new RegExp(
  "^\\s*(?:[A-Z0-9]{1,9})?\\s*$"
)

export const OpendartNullableStockCode = /*#__PURE__*/ zod.union([
  /*#__PURE__*/ zod
    .string()
    .check(/*#__PURE__*/ zod.regex(opendartNullableStockCodeOneRegExp)),
  /*#__PURE__*/ zod.null(),
])

export type OpendartNullableStockCode = zod.input<
  typeof OpendartNullableStockCode
>
export type OpendartNullableStockCodeOutput = zod.output<
  typeof OpendartNullableStockCode
>

export const OpendartCompanyResponse = /*#__PURE__*/ zod.object({
  status: /*#__PURE__*/ zod.literal("000"),
  message: /*#__PURE__*/ zod.optional(/*#__PURE__*/ zod.string()),
  corp_code: OpendartCorpCode,
  corp_name: /*#__PURE__*/ zod.string().check(/*#__PURE__*/ zod.minLength(1)),
  stock_code: /*#__PURE__*/ zod.optional(OpendartNullableStockCode),
  stock_name: /*#__PURE__*/ zod.nullish(/*#__PURE__*/ zod.string()),
  corp_cls: /*#__PURE__*/ zod.string().check(/*#__PURE__*/ zod.minLength(1)),
  induty_code: /*#__PURE__*/ zod.nullish(/*#__PURE__*/ zod.string()),
  est_dt: /*#__PURE__*/ zod.nullish(/*#__PURE__*/ zod.string()),
  acc_mt: /*#__PURE__*/ zod.nullish(/*#__PURE__*/ zod.string()),
})

export type OpendartCompanyResponse = zod.input<typeof OpendartCompanyResponse>
export type OpendartCompanyResponseOutput = zod.output<
  typeof OpendartCompanyResponse
>

export const opendartDisclosureRceptDtRegExp = new RegExp("^[0-9]{8}$")
export const opendartDisclosureRmDefault = ``

export const OpendartDisclosure = /*#__PURE__*/ zod.object({
  rcept_no: /*#__PURE__*/ zod.string().check(/*#__PURE__*/ zod.minLength(1)),
  corp_code: OpendartCorpCode,
  stock_code: /*#__PURE__*/ zod.optional(OpendartNullableStockCode),
  corp_name: /*#__PURE__*/ zod.string().check(/*#__PURE__*/ zod.minLength(1)),
  report_nm: /*#__PURE__*/ zod.string().check(/*#__PURE__*/ zod.minLength(1)),
  flr_nm: /*#__PURE__*/ zod.string().check(/*#__PURE__*/ zod.minLength(1)),
  rcept_dt: /*#__PURE__*/ zod
    .string()
    .check(/*#__PURE__*/ zod.regex(opendartDisclosureRceptDtRegExp)),
  rm: /*#__PURE__*/ zod._default(
    /*#__PURE__*/ zod.string(),
    opendartDisclosureRmDefault
  ),
})

export type OpendartDisclosure = zod.input<typeof OpendartDisclosure>
export type OpendartDisclosureOutput = zod.output<typeof OpendartDisclosure>

export const OpendartDisclosuresResponse = /*#__PURE__*/ zod.object({
  status: /*#__PURE__*/ zod.literal("000"),
  message: /*#__PURE__*/ zod.optional(/*#__PURE__*/ zod.string()),
  list: /*#__PURE__*/ zod.optional(
    /*#__PURE__*/ zod.union([
      /*#__PURE__*/ zod.array(OpendartDisclosure),
      OpendartDisclosure,
    ])
  ),
})

export type OpendartDisclosuresResponse = zod.input<
  typeof OpendartDisclosuresResponse
>
export type OpendartDisclosuresResponseOutput = zod.output<
  typeof OpendartDisclosuresResponse
>

export const opendartFinancialAccountBsnsYearRegExp = new RegExp("^[0-9]{4}$")

export const OpendartFinancialAccount = /*#__PURE__*/ zod.object({
  bsns_year: /*#__PURE__*/ zod
    .string()
    .check(/*#__PURE__*/ zod.regex(opendartFinancialAccountBsnsYearRegExp)),
  reprt_code: /*#__PURE__*/ zod.enum(["11013", "11012", "11014", "11011"]),
  corp_code: OpendartCorpCode,
  sj_div: /*#__PURE__*/ zod.string().check(/*#__PURE__*/ zod.minLength(1)),
  account_id: /*#__PURE__*/ zod.nullish(/*#__PURE__*/ zod.string()),
  account_nm: /*#__PURE__*/ zod.string().check(/*#__PURE__*/ zod.minLength(1)),
  thstrm_amount: /*#__PURE__*/ zod.nullish(
    /*#__PURE__*/ zod.union([
      /*#__PURE__*/ zod.string(),
      /*#__PURE__*/ zod.number(),
    ])
  ),
  frmtrm_amount: /*#__PURE__*/ zod.nullish(
    /*#__PURE__*/ zod.union([
      /*#__PURE__*/ zod.string(),
      /*#__PURE__*/ zod.number(),
    ])
  ),
  bfefrmtrm_amount: /*#__PURE__*/ zod.nullish(
    /*#__PURE__*/ zod.union([
      /*#__PURE__*/ zod.string(),
      /*#__PURE__*/ zod.number(),
    ])
  ),
  currency: /*#__PURE__*/ zod.nullish(/*#__PURE__*/ zod.string()),
})

export type OpendartFinancialAccount = zod.input<
  typeof OpendartFinancialAccount
>
export type OpendartFinancialAccountOutput = zod.output<
  typeof OpendartFinancialAccount
>

export const OpendartFinancialAccountsResponse = /*#__PURE__*/ zod.object({
  status: /*#__PURE__*/ zod.literal("000"),
  message: /*#__PURE__*/ zod.optional(/*#__PURE__*/ zod.string()),
  list: /*#__PURE__*/ zod.optional(
    /*#__PURE__*/ zod.union([
      /*#__PURE__*/ zod.array(OpendartFinancialAccount),
      OpendartFinancialAccount,
    ])
  ),
})

export type OpendartFinancialAccountsResponse = zod.input<
  typeof OpendartFinancialAccountsResponse
>